services/bot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==0.19.0
asyncpg==0.29.0
aiogram==2.25.1
aiohttp==3.8.5 --pre
python-multipart==0.0.6
certifi==2023.11.17
pytz==2024.1
redis>=5.0.0
hiredis>=2.0.0
aiokafka==0.7.2
kafka-python==2.0.2
confluent-kafka==2.3.0 